refactor: certain constants should be configs
